refactoring fetchDoggy for codeclimate improvement and filelist composable
