+++ src/helpers/FetchDoggy.js
@@ -24,6 +24,48 @@
  *
  * N.b. these properties are normal public properties so that Vue can make them reactive
  */
+async function readResponse(responseReader, callHooks, controller) {
+  let done = false;
+  let value;
+  while (!done) {
+    ({ value, done } = await responseReader.read());
+    if (done) break;
+    callHooks("progress", value.byteLength);
+    controller.enqueue(value);
+  }
+  callHooks("complete");
+  controller.close();
+}
+function responseGenerator(responseReader, callHooks, signal) {
+  return new Response(
+    new ReadableStream({
+      async start(controller) {
+        if (signal.aborted) {
+          controller.close();
+          await responseReader.cancel();
+        }
+        await readResponse(responseReader, callHooks, controller).catch((error) => {
+          controller(error);
+          callHooks("error");
+        });
+      },
+    })
+  );
+}
+
+function testResponse(response) {
+  if (!response.body) {
+    throw Error("ReadableStream is not yet supported in this browser.");
+  }
+
+  if (!response.ok) {
+    // HTTP error server response
+    const error = Error(`Server responded ${response.status} ${response.statusText}`);
+    this.callHooks("error", error);
+    throw error;
+  }
+}
+
 export default class FetchDoggy {
   constructor() {
     this.progress = 0;
@@ -36,6 +78,7 @@ export default class FetchDoggy {
       cancel: {},
       error: {},
     };
+    this.onProgress((amount) => (this.progress += amount));
   }
 
   // mimic native fetch() instantiation and return Promise
@@ -44,67 +87,27 @@ export default class FetchDoggy {
     const { signal } = this.controller;
     return fetch(request, { signal, ...init })
       .then((response) => {
-        if (!response.body) {
-          throw Error("ReadableStream is not yet supported in this browser.");
-        }
-
-        if (!response.ok) {
-          // HTTP error server response
-          const error = Error(`Server responded ${response.status} ${response.statusText}`);
-          this.callHooks("error", error);
-          throw error;
-        }
-
+        testResponse(response);
         // to access headers, server must send CORS header
         // "Access-Control-Expose-Headers: content-encoding, content-length x-file-size"
         // server must send custom x-file-size header if gzip or other content-encoding is used
-        const contentLength = response.headers.get(
-          response.headers.get("content-encoding") ? "x-file-size" : "content-length"
+        this.total = parseInt(
+          response.headers.get(
+            response.headers.get("content-encoding") ? "x-file-size" : "content-length"
+          ),
+          10
         );
 
         // don't evaluate download progress if we can't compare against a total size
-        if (contentLength === null) throw Error("Response size header unavailable");
-
-        this.total = parseInt(contentLength, 10);
-        this.onProgress((amount) => {
-          this.progress += amount;
-        });
+        if (isNaN(this.total)) throw Error("Response size header unavailable");
 
         // ensure onProgress called when content-length=0
         if (this.total === 0) {
-          this.callHooks("progress");
+          this.callHooks("progress", 0);
           this.callHooks("complete");
         }
 
-        const responseReader = response.body.getReader();
-        const callHooks = this.callHooks.bind(this);
-
-        return new Response(
-          new ReadableStream({
-            async start(controller) {
-              if (signal.aborted) {
-                controller.close();
-                await responseReader.cancel();
-              }
-              async function read() {
-                let done = false;
-                let value;
-                while (!done) {
-                  ({ value, done } = await responseReader.read());
-                  if (done) break;
-                  callHooks("progress", value.byteLength);
-                  controller.enqueue(value);
-                }
-                callHooks("complete");
-                controller.close();
-              }
-              await read().catch((error) => {
-                controller(error);
-                callHooks("error");
-              });
-            },
-          })
-        );
+        return responseGenerator(response.body.getReader(), this.callHooks.bind(this), signal);
       })
       .then((response) => response.arrayBuffer())
       .then((arrayBuffer) => new Blob([arrayBuffer], { type: extraOptions.mimetype }))
